requirements.txt
Flask
requests
numpy
sentence-transformers
pyyaml
gunicorn
Werkzeug
huggingface-hub
